feat(preset): loadLynxConfig + sink config-loading (loadConfig/validate/register/typia) into preset (B: auto-load lynx.config.ts)
