+++ examples/react-rsbuild/rsbuild.config.ts
@@ -4,13 +4,19 @@
 
 import { defineConfig } from '@rsbuild/core';
 
-import { pluginLynxPreset } from '@lynx-js/preset-rsbuild-plugin';
+import {
+  loadLynxConfig,
+  pluginLynxPreset,
+} from '@lynx-js/preset-rsbuild-plugin';
 import { pluginQRCode } from '@lynx-js/qrcode-rsbuild-plugin';
 import { pluginReactLynx } from '@lynx-js/react-rsbuild-plugin';
 
-export default defineConfig({
+// Progressive migration: the existing `lynx.config.ts` is loaded as-is and fed
+// to `pluginLynxPreset()`. `defineConfig` accepts an async factory, so the
+// config can be loaded before the plugins are composed.
+export default defineConfig(async () => ({
   plugins: [
-    pluginLynxPreset(),
+    pluginLynxPreset(await loadLynxConfig()),
     pluginReactLynx(),
     pluginQRCode({
       fullscreen: true,
@@ -20,10 +26,4 @@ export default defineConfig({
     web: {},
     lynx: {},
   },
-  source: {
-    // Align with the Rspeedy default entry name (`main`).
-    entry: {
-      main: './src/index.tsx',
-    },
-  },
-});
+}));
