+++ examples/react-rsbuild/tsconfig.json
@@ -9,7 +9,7 @@
     "checkJs": true,
     "isolatedDeclarations": false,
   },
-  "include": ["src", "rsbuild.config.ts"],
+  "include": ["src", "rsbuild.config.ts", "lynx.config.ts"],
   "references": [
     { "path": "../../packages/react/tsconfig.json" },
     { "path": "../../packages/rspeedy/core/tsconfig.build.json" },
