+++ packages/rspeedy/core/package.json
@@ -19,12 +19,6 @@
     "email": "colinwang.0616@gmail.com"
   },
   "type": "module",
-  "imports": {
-    "#register": {
-      "types": "./register/index.d.ts",
-      "import": "./register/index.js"
-    }
-  },
   "exports": {
     ".": {
       "types": "./lib/index.d.ts",
@@ -82,10 +76,7 @@
     "rsbuild-plugin-arethetypeswrong": "0.3.1",
     "rsbuild-plugin-publint": "0.3.4",
     "tiny-invariant": "^1.3.3",
-    "ts-blank-space": "^0.7.0",
-    "type-fest": "^5.4.4",
-    "typia": "12.1.1",
-    "typia-rspack-plugin": "3.0.1"
+    "type-fest": "^5.4.4"
   },
   "peerDependencies": {
     "typescript": "5.1.6 - 5.9.x"
@@ -112,12 +103,6 @@
         "types": "./client.d.ts"
       },
       "./package.json": "./package.json"
-    },
-    "imports": {
-      "#register": {
-        "types": "./dist/register/index.d.ts",
-        "import": "./dist/register/index.js"
-      }
     },
     "types": "./dist/index.d.ts"
   },
