+++ packages/rspeedy/core/test/cli/fixtures/register.js
@@ -1,3 +1,4 @@
-import { register } from '../../../register/index.js'
+// The TS-loader `register` moved to `@lynx-js/preset-rsbuild-plugin`.
+import { register } from '../../../../plugin-preset/register/index.js'
 
 register()
