+++ packages/rspeedy/core/tsconfig.build.json
@@ -1,11 +1,6 @@
 {
   "extends": "../../../tsconfig.json",
   "compilerOptions": {
-    "plugins": [
-      {
-        "transform": "typia/lib/transform",
-      },
-    ],
     "strictNullChecks": true,
     "strict": true,
     "composite": true,
@@ -19,7 +14,6 @@
     "resolveJsonModule": true,
   },
   "include": [
-    "register",
     "src",
   ],
   "references": [
