+++ packages/rspeedy/core/tsconfig.json
@@ -6,7 +6,6 @@
   },
   "include": [
     "bin",
-    "register",
     "src",
     "test",
     "client.d.ts",
