+++ packages/rspeedy/plugin-preset/register/data.d.ts
