+++ packages/rspeedy/plugin-preset/register/hooks.js
