+++ packages/rspeedy/plugin-preset/register/index.d.ts
