+++ packages/rspeedy/plugin-preset/register/index.js
