+++ packages/rspeedy/plugin-preset/src/config/defineConfig.ts
@@ -2,7 +2,7 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 
-import type { Config } from '@lynx-js/preset-rsbuild-plugin'
+import type { Config } from './index.js'
 
 /**
  *  Parameters for the function exported from `lynx.config.js`.
