+++ packages/rspeedy/plugin-preset/test/config/fixtures/cjs/default/config.mjs
