+++ packages/rspeedy/plugin-preset/test/config/fixtures/cjs/default/package.json
