+++ packages/rspeedy/plugin-preset/test/config/fixtures/cjs/esm-import-esm.js
