+++ packages/rspeedy/plugin-preset/test/config/fixtures/cjs/export-default.ts
