+++ packages/rspeedy/plugin-preset/test/config/fixtures/cjs/lynx.config.js
