+++ packages/rspeedy/plugin-preset/test/config/fixtures/cjs/package.json
