+++ packages/rspeedy/plugin-preset/test/config/fixtures/custom/custom-cts.config.ts
