+++ packages/rspeedy/plugin-preset/test/config/fixtures/custom/custom.config.cjs
