+++ packages/rspeedy/plugin-preset/test/config/fixtures/custom/enum.ts
