+++ packages/rspeedy/plugin-preset/test/config/fixtures/custom/function-promise.ts
