+++ packages/rspeedy/plugin-preset/test/config/fixtures/custom/function.ts
