+++ packages/rspeedy/plugin-preset/test/config/fixtures/custom/promise.ts
