+++ packages/rspeedy/plugin-preset/test/config/fixtures/error/invalid.js
