+++ packages/rspeedy/plugin-preset/test/config/fixtures/error/throw.js
