+++ packages/rspeedy/plugin-preset/test/config/loadConfig.test.ts
