+++ packages/rspeedy/plugin-preset/test/config/validate.test.ts
