+++ packages/rspeedy/plugin-preset/tsconfig.build.json
@@ -1,6 +1,11 @@
 {
   "extends": "../../../tsconfig.json",
   "compilerOptions": {
+    "plugins": [
+      {
+        "transform": "typia/lib/transform",
+      },
+    ],
     "composite": true,
     "outDir": "./lib",
     "rootDir": "./src",
