+++ packages/rspeedy/plugin-preset/tsconfig.json
@@ -4,7 +4,7 @@
     "rootDir": ".",
     "noEmit": true,
   },
-  "include": ["src", "test", "rstest.config.ts", "package.json"],
+  "include": ["src", "register", "test", "rstest.config.ts", "package.json"],
   "exclude": ["test/**/fixtures"],
   "references": [
     { "path": "../plugin-debug-metadata/tsconfig.build.json" },
